Fast and Slow pointer working.
Here, I have understood how the fast and slow pointer(Tortoise and hare algorithm) is working, 
First this algorithm was used to found the center of a LL, cycling loops in LL.
This works like(for finding the center of the node), 
Two variables(slow and fast) point to the same node initially, then we start moving those pointers as slow pointer by one step and fast pointer by two step, 
And imagine the size of the list is 10, then when the fast pointer reaches the final node(that is the 10th node), the slow pointer will be automatically at the half of 10, that is 10/2 is 5. Which is the center of the node. 
So for every 1 step of the slow pointer fast pointer will move 2 steps.

The working of the cycle detection,
Consider there are 13 nodes in LL, and the questions criteria is there must be a cycle in that 13 nodes,  
so if we start with two pointer slow and fast, when the slow pointer reaches the 13th node, the fast pointer have gone for 2 x 13 = 26 nodes, so when the slow pointer points exactly to the last node(slow meets this node for the first time), at that time the fast pointer will also points exactly at the last node(fast meets this point for teh second time).